chore: update packages scope from `@bee-q` to `@beeq` (#863)
This PR aligns the package scope name with the current Design System name. Projects using the old package scope should update their dependencies accordingly, depending on the package used, e.g: `npm i @beeq/{core,react}`

Since this only changes the NPM scope of each package, we decided to target it as a path version rather than release a new major version only because of that. We are human, and sometimes mistakes happen, so we apologize for any trouble and inconvenience this could cause you.